feat: add test infrastructure for next app
