+++ package.json
@@ -68,6 +68,7 @@
     "remark-admonitions": "1.2.1",
     "remarkable": "2.0.1",
     "text-runner": "^5.0.0",
+    "wait-on": "^6.0.0",
     "widdershins": "4.0.1",
     "yaml": "1.10.2"
   }
